Handle clangWarnings as errors, handle script and linker errors
